+++ pkg/helpers/helpers_test.go
@@ -185,7 +185,3 @@ func (i unsupportedObjectIdentifier) Name() string {
 func (i unsupportedObjectIdentifier) FullyQualifiedName() string {
 	return "fully qualified name"
 }
-
-func (i unsupportedObjectIdentifier) Representation() string {
-	return ""
-}
