+++ pkg/resources/helpers_test.go
@@ -3,9 +3,6 @@ package resources_test
 import (
 	"testing"
 
-	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/sdk"
-	"github.com/hashicorp/go-cty/cty"
-
 	"github.com/stretchr/testify/assert"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
@@ -472,138 +469,4 @@ func tagGrant(t *testing.T, id string, params map[string]interface{}) *schema.Re
 	r.NotNil(d)
 	d.SetId(id)
 	return d
-}
-
-func TestIsDataType(t *testing.T) {
-	isDataType := resources.IsDataType()
-	key := "tag"
-
-	testCases := []struct {
-		Name  string
-		Value any
-		Error string
-	}{
-		{
-			Name:  "validation: correct DataType value",
-			Value: "NUMBER",
-		},
-		{
-			Name:  "validation: correct DataType value in lowercase",
-			Value: "number",
-		},
-		{
-			Name:  "validation: incorrect DataType value",
-			Value: "invalid data type",
-			Error: "expected tag to be one of",
-		},
-		{
-			Name:  "validation: incorrect value type",
-			Value: 123,
-			Error: "expected type of tag to be string",
-		},
-	}
-
-	for _, tt := range testCases {
-		t.Run(tt.Name, func(t *testing.T) {
-			_, errors := isDataType(tt.Value, key)
-			if tt.Error != "" {
-				assert.Len(t, errors, 1)
-				assert.ErrorContains(t, errors[0], tt.Error)
-			} else {
-				assert.Len(t, errors, 0)
-			}
-		})
-	}
-}
-
-func TestIsValidIdentifier(t *testing.T) {
-	accountObjectIdentifierCheck := resources.IsValidIdentifier[sdk.AccountObjectIdentifier]()
-	databaseObjectIdentifierCheck := resources.IsValidIdentifier[sdk.DatabaseObjectIdentifier]()
-	schemaObjectIdentifierCheck := resources.IsValidIdentifier[sdk.SchemaObjectIdentifier]()
-	externalObjectIdentifierCheck := resources.IsValidIdentifier[sdk.ExternalObjectIdentifier]()
-	tableColumnIdentifierCheck := resources.IsValidIdentifier[sdk.TableColumnIdentifier]()
-
-	testCases := []struct {
-		Name       string
-		Value      any
-		Error      string
-		CheckingFn schema.SchemaValidateDiagFunc
-	}{
-		{
-			Name:       "validation: invalid value type",
-			Value:      123,
-			Error:      "Expected schema string type, but got: int",
-			CheckingFn: accountObjectIdentifierCheck,
-		},
-		{
-			Name:       "validation: invalid identifier representation",
-			Value:      "",
-			Error:      "Unable to parse the identifier: ",
-			CheckingFn: accountObjectIdentifierCheck,
-		},
-		// Invalid form for different checkers (tests getExpectedIdentifierForm function)
-		{
-			Name:       "validation: incorrect form for account object identifier",
-			Value:      "a.b",
-			Error:      "<name>, but was <database_name>.<name>",
-			CheckingFn: accountObjectIdentifierCheck,
-		},
-		{
-			Name:       "validation: incorrect form for database object identifier",
-			Value:      "a.b.c",
-			Error:      "<database_name>.<name>, but was <database_name>.<schema_name>.<name>",
-			CheckingFn: databaseObjectIdentifierCheck,
-		},
-		{
-			Name:       "validation: incorrect form for schema object identifier",
-			Value:      "a.b.c.d",
-			Error:      "<database_name>.<schema_name>.<name>, but was <database_name>.<schema_name>.<table_name>.<column_name>",
-			CheckingFn: schemaObjectIdentifierCheck,
-		},
-		{
-			Name:       "validation: incorrect form for table column identifier",
-			Value:      "a",
-			Error:      "<database_name>.<schema_name>.<table_name>.<column_name>, but was <name>",
-			CheckingFn: tableColumnIdentifierCheck,
-		},
-		{
-			Name:       "validation: external object identifier is validated",
-			Value:      "a",
-			Error:      "Identifier validation is not available for sdk.ExternalObjectIdentifier type.",
-			CheckingFn: externalObjectIdentifierCheck,
-		},
-		// Valid form for different checkers (tests getExpectedIdentifierForm function)
-		{
-			Name:       "correct form for account object identifier",
-			Value:      "a",
-			CheckingFn: accountObjectIdentifierCheck,
-		},
-		{
-			Name:       "correct form for database object identifier",
-			Value:      "a.b",
-			CheckingFn: databaseObjectIdentifierCheck,
-		},
-		{
-			Name:       "correct form for schema object identifier",
-			Value:      "a.b.c",
-			CheckingFn: schemaObjectIdentifierCheck,
-		},
-		{
-			Name:       "correct form for table column identifier",
-			Value:      "a.b.c.d",
-			CheckingFn: tableColumnIdentifierCheck,
-		},
-	}
-
-	for _, tt := range testCases {
-		t.Run(tt.Name, func(t *testing.T) {
-			diag := tt.CheckingFn(tt.Value, cty.IndexStringPath("path"))
-			if tt.Error != "" {
-				assert.Len(t, diag, 1)
-				assert.Contains(t, diag[0].Detail, tt.Error)
-			} else {
-				assert.Len(t, diag, 0)
-			}
-		})
-	}
 }
